+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (@egomobile/http-supertest)
 
+## 0.4.0
+
+- improve output of test progress
+- add `getStream` and `groupPrefix` props to [ISetupTestEventListenerOptions interface](https://egomobile.github.io/node-http-supertest/interfaces/ISetupTestEventListenerOptions.html)
+- update to [@egomobile/http-server](https://github.com/egomobile/node-http-server) `^0.58.0`
+
 ## 0.3.1
 
 - update to [@egomobile/http-server](https://github.com/egomobile/node-http-server) `^0.57.0`